plugin: add qos member to bank_info struct
Add a new member to the bank_info struct: qos, a
vector of strings which will hold all available
Quality of Services to a particular user/bank row.
The QoS' are passed in as a comma-delimited string,
and then parsed and pushed one-by-one into the qos
member in the bank_info struct.