plugin: add QoS validation
Add validation for an optional QoS argument
when a job is submitted. The QoS is first checked
to exist in the qos_map. It is then checked to
determine if it is a valid QoS for a user to specify
when submitting their job. If a valid QoS is
passed in, it is added to to the user's bank_info
struct.