t1000: move and rename expected files
Move the test DB's and expected files for t1000-print_hierarchy.t
to their own directory within t/expected/, called print_hierarchy/.
Rename the expected files to have the ".expected" suffix and rename
the test files to have the ".test" suffix. This is consistent with
the other sharness tests that also compare test and expected files.