t1011: fix grep command for post_usage .txt files
Problem: The "grep" command is being used incorrectly in the sharness
test t1011-job-archive-interface.t when looking for certain values in
the output of "flux account-shares," which results in test failures.

Move the name of the file being grep-ed to the end of the command, which
is consistent with its other uses throughout the other sharness tests
in this repository.