plugin: add jobspec-update of project name
Problem: If a user submits a job under a default project, there is no
way to tell which project they submitted their job under because it does
not get posted to jobspec. This has two cases:

1) a user/bank with information already loaded to the plugin submits a
job under a default project:

Post the project name in a jobspec-update during job.validate by calling
the update_jobspec_project () helper function in the case where a
user/bank does NOT specify a project when submitting their job.

2) a user/bank with information NOT loaded to the plugin submits a job
under a default project:

Post the project name by calling the update_jobspec_project () helper
function in the section of job.state.priority where the job is being
updated with the user/bank information.