create_db: add "jobs" and "last_seen_job" tables
Add a new table to the flux-accounting DB: jobs, which will store jobs
fetched by flux-accounting periodically to be used for job-usage and
fair-share calculation as well as for viewing by users/admins.

Add a new last_seen_job table to the flux-accounting DB, which will
hold a timestamp of the most recently seen inactive job. The
fetch-job-records.py script will use this timestamp as a filter when
looking for newly completed jobs (i.e look for jobs that have completed
after this timestamp).