t1011: remove ntasks count from job submit
Problem: Specifying an ntasks count in the job submit commands
in t1011-job-archive-interface.t causes more nodes to be used than
intended on a 2021 Macbook Pro, which will result in failed tests since
the test looks for a certain number of nodes per user.

Since the test is really only concerned with getting an accurate number
of nodes per user when looking at flux account-shares, just remove the
specification of ntasks from the job submit commands that use it, and
only specify nodes in the job submissions.