docker: add arm64 image to testsuite
Problem: flux-accounting doesn't have an arm64 Docker image to build
and interact with, and on my Macbook Pro (which I believe is an arm
based machine), the other Docker containers (el8 and jammy) are very
slow and not fun to develop in.

Add an arm64 Docker image to the testsuite.