docker-run-checks.sh: add "platform" argument
Problem: The docker-run-checks.sh script has no way to specify which
platform to run the container on. For arm-based containers, an
arm-based platform needs to be specified in order for the container to
be built and run.

Copy flux-core's docker-run-checks.sh implementation of the "platform"
argument and add it to flux-accounting's docker-run-checks.sh script.