plugin: change char * variables to const char *
Problem: The multi-factor priority plugin unpacks string variables and
stores them in a char *, which is a mutable pointer to a mutable string.
However, there are cases where  we don't want the contents of the
variable it is pointing to to change.

Change the types of these char * variables to const char *.