bank_info: add get_user_info () helper function
Problem: There is some functionality in the priority plugin
(mf_priority.cpp) that is specific to accessing user_bank_info objects
in a map data structure. These methods can be abstracted out to the
class files to clean up some of the plugin code.

Add a new function to the bank_info.cpp class file that handles
looking up a user/bank combo and fetching a user_bank_info object from
the users map, or returning NULL on failure.