plugin: change args to get_queue_info ()
Problem: The arguments to the get_queue_info () function include an
iterator to the bank_info object, which is used in the function to
access the user/bank's list of permissible queues for validation. This
can be confusing to read and would be easier if just the list of
permissible queues was passed to the function.

Change the argument to the get_queue_info () function from an iterator
to the user/bank object to just the user/bank's list of permissible
queues.

Change the calls to this function to account for the argument change.