plugin: improve priority_cb ()
Problem: The callback for job.state.priority does not make use of the
new external functions in bank_info.cpp that can look up a given
user/bank, and instead employs its own methods of performing this
lookup, creating a bit of duplicate code and more complex readability.

Improve the callback for job.state.priority by making use of the new
external user/bank lookup function(s) in bank_info.cpp.

Add additional comments in the callback, particularly in the cases where
a job was submitted before the plugin was loaded with flux-accounting
data and needs to update the current job with the correct accounting
information.