t: edit expected output of exception msg
Problem: There is a test in t1012-mf-priority-load.t that checks the
contents of the exception message that is raised when a user/bank cannot
be validated when their previously held job is being updated as a result
of the priority plugin loading flux-accounting information. However,
the job.state.priority callback has been updated and a new exception
message is returned, so this test fails.

Update the expected output of the exception message that is tested in
t1012-mf-priority-load.t to match the one that is returned.