plugin: improve callback for job.state.depend
Problem: The variable name in the callback for job.state.depend is sort
of poorly named. The callback also does not have a comment describing
its purpose.

Change the variable name for the pointer to user/bank information in the
callback for job.state.depend.

Add a comment to the function describing its purpose.