plugin: expand split_string ()
Problem: The split_string () helper function can only accept a list of
queues to parse when unpacking user/bank information from the
flux-accounting DB, but there also exists a need to be able to parse a
list of projects. Since both lists can be parsed the same way,
this helper function should be expanded.

Expand the split_string () helper function to accept a list of strings
to parse through when unpacking user/bank information. This allows it to
unpack either a list of queues or a list of projects.