t: move sample JSON payloads to separate files
Problem: The sample user payloads sent to the plugin grow longer than
wanted in the sharness test files with the addition of the "project" and
"default_project" fields.

For the tests that have these sample payloads, create a new folder in
the t/ directory that stores these JSON payloads permanently in a file
instead of creating them within the test.