plugin: move user/bank insert, split_str ()
Problem 1: The priority plugin contains code in the callback that's
responsible for unpacking user/bank information that directly
manipulates the users map by inserting user/bank information in it.
This functionality can be abstracted to a helper function in the
external files that deal directly with user/bank information.

Problem 2: The split_str_push_back () helper function is really only
defined to help with the insertion of user/bank information in the
users map, but it is located directly within the plugin code. This
function can also be moved to the external files taht deal directly
with user/bank information.

Add a new helper function called insert_user () that inserts info about
user/banks into the users map.

Move the split_str_push_back () function to an external file and change
its arguments to be more C++-oriented and accept a const std::string&
instead of a char*.

As a result of the parameter change to split_str_push_back (), the
expected output of the plugin.query callback is also improved; in the
case where a user/bank does not belong to any queues, their "queues"
field used to show an empty set of quotes (""). Now, it just shows an
empty list. Modify the expected output of the plugin states so that the
tests in t1019-mf-priority-info-fetch.t pass.