job.updated: improve callback(s)
Problem: There are now helper functions available for the plugin to make
use of to help with lookups and access of the internal map for users and
banks, but the plugin does not make use of them.

Restructure the job_updated () and update_queue_cb () callbacks to make
use of the new get_user_info () helper function.

Improve the comments in these functions/add new ones to help with some
of the clarity in each function.