accounting: add get_queue_info () helper function
Problem: There is a function in the multi-factor priority plugin that
validates a queue for an Association when they submit a job. This is
specific to flux-accounting, and as such it should be moved out of the
plugin and moved to the extrnal files that hold Accounting stuff.

Add the get_queue_info () helper function to the accounting.* files. Add
the queues map as a parameter to this function.