view-job-records: fail gracefully when no DB found
Problem: view-job-records does not assume a default path for the
flux-core job-archive DB when a path is not specified and it results in
a very long stack trace.

Second problem: If a connection to a DB fails in
establish_sqlite_connection(), the function will try to call sys.exit(1)
which will result in a hang.

Change this behavior to instead return -1. Check for an error code when
calling this function from view_job_records () and raise an Exception if
the connection fails.

Add a test to t1011-job-archive-interface.t to check for this Exception
when a connection to a DB fails.