plugin: remove check_queue_factor () helper func
Problem: The check_queue_factor () helper function is just checking the
result of the get_queue_info () function for INVALID_QUEUE. This could
easily be just done in-line whenever the check_queue_factor () function
is called.

Remove the definition of check_queue_factor () and instead just check
for the result of get_queue_info () manually in the place where it is
called.