job_archive_interface: don't add canceled jobs
Problem: When a job does not run or is canceled, the "R" field in that
job will just be empty quotes (''), which causes an error when trying to
assign that field to a ResourceSet object. Furthermore, it should not be
considered when calculating the job-usage value for an association.

Add a check in job_archive_interface.py that checks for the contents of
the "R" field before trying to assign it to a ResourceSet object. If it
is empty, skip adding this job to the association's list of jobs.

Add some tests to t1011-job-archive-interface.t that test submitting a
job that never runs and ensuring that it does not affect the
association's job usage value as well as does not show up in their jobs
with the view-job-records command.