database: add 'max_cores' column
Problem: The flux-accounting database has no column to store the max
number of cores an association can have across all of their running
jobs.

Add a new 'max_cores' column to the 'association_table'. Update the
Python bindings and flux-accounting service to account for the new
column, notably the 'add-user' and 'view-user' functions.

Update the schema version of the database as a result of the new column
addition.