plugin: add 'max_cores' to information update
Problem: An association's max-cores number is not sent to the priority
plugin, which will ultimately be responsible for enforcing this limit
across an asssociation's set of running jobs.

Add 'max-cores' to the set of information sent from the flux-accounting
DB to the priority plugin.

Add 'max-cores' as a member to the Association class in accounting.hpp.

Add unpack logic in the plugin to include max-cores information. Add a
default assignment for max-cores in the case where a special temporary
Association object needs to be created.

Add 'max-cores' to the JSON object created for each association when
constructing the map returned in the plugin.query callback.