accounting: add "cur_nodes" as a member
Problem: The Association class has no member to track an association's
current node count across all of their running jobs.

Add a new member called "cur_nodes" which represents an association's
current node count across all of their running jobs.

Add a default value in the case where we are creating a special
temporary association while the plugin waits for flux-accounting data
to be loaded in.

Add cur_node values to the accounting unit tests as well as to the
expected output of the plugin.query tests.