script: update keys, add eventlog
Add "eventlog" to the set of information fetched for each job.

Edit the "nodelists" key to be "nodelist" so that it actually grabs
the nodelist for each job.

Compute eligible time for each job using t_depend and t_run and add it
to each job dictionary.

Compute proc.count for each job using nnodes and ntasks and add it to
each job dictionary.