python: add new util folder
Problem: flux-accounting has lots of different tables that can be viewed
with view-* and list-* commands, but each one has its own function that
could probably be cleaned up with some common utility.

Add a new folder to the python bindings which will store common code for
the subcommands of flux-accounting. Add a util.py file which has
functions for outputting data from a table in the flux-accounting DB in
either table format or JSON format.