list-banks: use None as default
Problem: The default value for the --fields optional argument in the
list-banks command is a comma-separated string of values that represent
the names of the columns in the bank_table, but the default is already
defined in the function that list-bank calls.

Change the default for the --fields optional argument to be None if it
is not specified.