cmd: add calc-project-usage command
Problem: Administrators have to manually calculate total job usage by
project by fetching all jobs under a certain project name using the
"view-job-records" command and then calculating it themselves. It would
be easier if flux-accounting did this for them.

Add a new command to the flux-accounting command suite that will
calculate job usage by project. If no project or timestamps are
specified, the command will calculate the all-time usage for each
registered project and update project_table with the usage. It also
takes optional arguments to specify a project name or certain
timestamps to further refine the usage calculated for a project. In
this case, the usage does *not* update project_table, and instead will
just be returned to stdout.