view-job-records: use new JobsFormatter class
Problem: The convert_to_str() function defines its own formatting for
job records to be returned with the view-job-records command, but the
formatter module in flux-accounting now has its own JobsFormatter class
that does this (and also includes custom formatting).

Edit the convert_to_str() function to create a JobsFormatter object,
initialized with the list of job_records to be returned. Add a --fields
optional argument to the view-job-records command to allow the user to
specify which fields to include in their output if they so choose.