t: edit expected output of view-job-records
Problem: The output of the view-job-records command has been slightly
adjusted as a result of including custom formatting in its output, but
there are a couple of tests throughout the testsuite that are now
out-of-date because they look for the old format.

Adjust the tests that look for view-job-records output throughout the
testsuite to account for the new formatting. Get rid of a expected file
in the testsuite in favor of just creating that file within the test
itself.