bindings: add load-config command
Problem: The flux-accounting DB has no way to load data from TOML
files, which is a common method of configuring other Flux components.
It would make sense if flux-accounting enabled the same capability,
especially with components that are shared across projects, such as
queues.

Add a new command to the command suite called load-config. It takes one
argument, a path to a TOML config file which is loaded and then parsed
for flux-accounting-specific properties. To start, just look for one
subsection called accounting.queue-priorities, which outlines queues
and their associated integer priorities. Iterate through this
subsection and add them to the queue_table in the flux-accounting DB.