plugin: track resources used per-job
Problem: The priority plugin does not keep track of the resources
allocated per-job when an association submits a job, which will be
needed when the plugin begins to enforce limits on how many resources
an association can use across all of their running jobs.

Add an increment/decrement of node+core counts in job.state.run and
job.state.inactive by unpacking the jobspec of the job and
calculating the number of nodes and cores assigned with the job.
Increment and decrement these values in the "cur_nodes" and "cur_cores"
attributes of the Association object.