plugin: add/rm dependencies due to resource usage
Problem: The priority plugin tracks resource usage across an
association's set of running jobs, but it does not enforce any limits
on this resource usage in the event where an association goes over
their limit.

Add a check in job.state.depend to see if the job's resources would put
the association over *either* their max_nodes or max_cores limit. If
so, add a dependency on the job and hold it.

In job.state.inactive, rework the check to see if there are any held
jobs for the association to also ensure the held job would satisfy
all three conditions:

1) the association is under their max-running-jobs limit
2) the job would not put the association over their max nodes limit
3) the job would not put the association over their max cores limit

If all three conditions are met, remove the dependency from the job.
Because there could be more than one dependency (the job is either held
due to a max running jobs limit *or* the job is held due to a max
resources limit), attempt to remove both dependencies.