plugin: add new Job class
Problem: As more flux-accounting dependencies are added to the priority
plugin, it will be harder to keep track of certain attributes of jobs
without having to unpack the jobspec for the job in multiple callbacks.

Create a new Job class for the priority plugin in order to keep track
of a held job's list of dependencies and its requested resource counts.
Add this class to its own source file and compile it along with the rest
of the priority plugin code.