edit-user: make "fairshare" an editable field
Problem: Since admins are the only ones who have privileges to run the
edit-user command, they should be able to edit the fairshare column for
an association manually if they so choose, but there is currently no
way to do so via flux-accounting command; they have to connect to a
SQLite shell and create and execute a SQL statement to update the
column manually.

Add the fairshare column as an editable field in the edit-user command.