pop-db: specify arguments to add_user() call
Problem: The pop-db command calls add_user() in order to add users from
a .csv file to the flux-acounting DB, but it does not specify which
args it is passing in the add_user() call. Now that the arguments to
add_user() have changed with the addition of the fairshare argument,
this call in pop-db breaks.

Specify which arguments are being passed to add_user() in pop-db.