data_reader_db: change shrs, usg types to uint64_t
Problem: The Account class used to build nodes for the weighted tree
when calculating fair-share values uses uint64_t for both the shares
and usage attributes for the class, but data_reader_db.cpp declares
these types as long long and double, respectively.

Change the types of these variables to uint64_t to match their expected
types in the Account class.