cmd: add toggle-wal-mode command
Problem: SQLite databases have the ability to enable Write-Ahead
Logging (WAL) mode, allowing concurrent reads and writes to the DB,
which would be a useful option to have set during large write
operations on the DB, such as a job-usage/fair-share update. This
setting is not currently toggle-able for the flux-accounting DB from
the command line, however.

Introduce a new command to the flux-accounting command suite called
toggle-wal-mode, which will toggle the WAL mode for the flux-accounting
DB on and off. Make it a privileged command so that only admins can run
this command.