cmd: add get-db-info command
Problem: The flux-accounting DB has useful metadata information that
can only be accessed by entering into a sqlite3 shell and writing
PRAGMA statements. It would be more convenient if flux-accounting
offered an option to view this information from the command-line.

Add a new command to the flux-accounting command suite called
get-db-info, which will return basic DB metadata such as the schema
version and whether Write-Ahead Logging (WAL) mode is enabled on the
DB.