queues: add load-queues command
Problem: flux-accounting doesn't have a way to initialize the
queue_table with queues already defined in a TOML file for the rest of
the flux components.

Add a load-queues command, which takes a filepath to a TOML file and
extracts all queue names defined with the "queues" prefix and calls
add_queue().