plugin: add unpack of bank information from DB
Problem: The priority plugin does not currently unpack bank attribute
information from the flux-accounting DB, but it will need to do this in
order to enforce/set a max-preempt-after attribute on a job if it is
running under a certain bank.

Add a section to the bulk update script to also send bank information,
particularly bank names and its max-preempt-after attribute to the
priority plugin.

Unpack the bank information in the plugin and store it in a map that
contains bank information.