accounting: add helper functions for dependencies
Problem: The checks for an association's max running jobs limits both
per-association and per-queue happen manually in the callback for
job.state.inactive. It would be cleaner to move this to helper
functions that the priority plugin could reference in the callback.

Add some helper functions to check an association's various limits both
per-association and per-queue.