Association: rm held_jobs, queue_held_jobs members
Problem: The Association class uses multiple members to keep track of
both held jobs in general and held jobs per-queue. Now that we have a
Job class, we can combine these into just one map, where the key is a
job ID, and the value is a Job objectm which will store data from both
of the previous structures, i.e its job ID as well as which queue it is
running in.

Remove the held_jobs and queue_held_jobs members from the Association
class in favor of using just one map, called hj (which stands for
"held jobs"). Adjust mf_priority.cpp as well as all unit tests to
account for the removal of held_jobs and queue_held_jobs and the
addition of the hj member.