queue_max_run_jobs: return max int instead of -1
Problem: The queue_max_run_jobs () function returns -1 if a queue
cannot be found in the plugin's internal map of known queues, to which
the priority plugin will just assign a max integer. It would be cleaner
if the helper function just returned the max integer itself instead of
the plugin having to check for -1 before assigning a max integer.

Change the return statement in queue_max_run_jobs () to return
std::numeric_limits<int>::max () instead of -1 when a queue cannot be
found.