plugin: use new Job class for dependencies
Problem: The priority plugin has to manually extract and check info
about both the job and the association that is submitting the job for
any potential flux-accounting limits in both job.state.depend and
job.state.inactive. As more and more dependencies are added to keep
track of in flux-accounting, it will get more cumbersome to have to do
these checks in the callbacks themselves.

Utilize the new Job class to do comparisons of an association's
flux-accounting limits when their job enters job.state.depend and
job.state.inactive. Move the names of the dependencies added to and
removed from jobs to constants in the plugin.