Association: rm held_jobs, queue_held_jobs members
Problem: The Association class uses multiple members to keep track of
both held jobs in general and held jobs per-queue. Now that we have a
Job class and one map that can keep track of all held jobs as well as
what queue they are running in, the held_jobs vector and
queue_held_jobs map are no longer needed as members.

Remove the held_jobs and queue_held_jobs members from the Association
class in favor of using just hj (which stands for "held jobs").
Adjust mf_priority.cpp as well as all tests to account for the removal
of held_jobs and queue_held_jobs and the addition of the hj member.