to_json (): add new hj map to JSON object
Problem: the new hj member of an Association object is not included in
the output of the to_json () method, which is used to report all
information that the plugin has about each Association in its internal
map. It should be included.

Add the hj map to the other information packed for each Association
object in the to_json () method.