plugin: add/rm dependencies due to resource usage
Problem: The priority plugin tracks resource usage across an
association's set of running jobs, but it does not enforce any limits
on this resource usage in the event where an association goes over
their limit.

Add a check in job.state.depend to see if the job's resources would put
the association over *either* their max_nodes or max_cores limit. If
so, add a dependency on the job and hold it.

In job.state.inactive, add a check to see if it would also satisfy the
association's max resources limit.