data_writer_db: wrap UPDATEs in transaction
Problem: data_writer_db () executes UPDATE statements for every
association in the flux-accounting DB to update their fair-share column.
If there are hundreds or thousands of users in the database, each UPDATE
statement will require a lock on the DB and cause other reads to the DB
(such as a "view-user" command) to hang until this process is done.

Wrap all of the UPDATE statements into a single transaction to all be
committed at once instead of committing one statement at a time.