cmd: add calc-job-priority command
Problem: There is no way for a user to see how an association's job
priority would be calculated given their current attributes without
consulting the flux-accounting documentation and plugging in values
manually in order to calculate it themselves.

Add a new command to the flux-accounting command suite that will
describe the calculation for an association's job's priority given their
current fairshare and any optional properties like a non-default bank
or queue (if a bank or queue are configured with a priority).

If priority factor weights are configured via TOML file, this file can
be passed in and used in the calculation of a job's priority. Otherwise,
default configuration weights are assumed when calculating a job's
priority.