update-usage: move half-life query
Problem: The calc_usage_factor() issues a query to fetch the timestamp
of the end of the current half-life period for every association when it
really only needs to perform this query once.

Move this query to before calculating the new job usage values for every
association and instead just pass it as a parameter to the
calc_usage_factor() function.

Adjust the unit tests in t/python/t1006_job_archive.py to account for
the new required argument to calc_usage_factor().