python: add basic graph capability
Problem: Some of the numerical data in the flux-accounting database
(such as job usage data) would be convenient to have compared across
rows in the association_table or bank_table without having to list the
tables themselves and parse data manually. It would be convenient to see
graphs where this data could be viewed visually.

Add new classes and subclasses to flux-accounting designed to create
horizontal bar charts for numerical data in flux-accounting. For now,
create subclasses for creating graphs to compare job usage values
between banks or associations.