queue_table: add "max_nodes" column
Problem: Queues do not have a property that represents how many nodes
an association can have across all of their running jobs in a certain
queue.

Add a new column to queue_table called "max_nodes", which is an
integer representing the max number of nodes an association can have
across all of their running jobs in a certain queue. Update the schema
version number in __init__.py.in to account for the addition of a new
column in queue_table.