queue commands: add --max-nodes optional arg
Problem: The max_nodes column for a row in queue_table cannot be
specified in either the add-queue or edit-queue commands.

Add max_nodes as an optional argument in the add-queue or edit-queue
commands. Edit the "min_nodes" and "max_nodes" parameters in add_queue()
to be more specific (since they are per-job properties).