Association: change queue_usage map structure
Problem: The queue_usage attribute of the Association class currently
only keeps track of the queue an association is running their jobs under
and how many jobs they have under a particular queue. However, there is
also a need to know how many resources (i.e. nodes) an association has
across all of their running jobs under a particular queue. So, the
plugin needs to be able to keep track of this number.

Change the queue_usage map attribute to be a map with a value of a new
QueueUsage class, a class that tracks an association's usage under a
particular queue. Give the QueueUsage class two attributes:
cur_run_jobs (the number of running jobs the association has in a
particular queue) and cur_nodes (the number of nodes an association has
in a particular queue). Update the way the cur_run_jobs value is
currently unpacked in the priority plugin to account for the QueueUsage
class.

Update the to_json() method for the Association class to pack these
QueueUsage objects in a JSON object. Update existing tests that check
the output of these JSON objects to account for the change in structure
of the JSON object.