dependencies: add per-queue max_nodes check
Problem: The plugin needs a way to enforce a per-queue max_nodes
dependency when an association is using up a lot of nodes across their
running jobs under a particular queue.

Begin to lay the groundwork for enforcing this limit. Define a constant
for the name of the dependency. Add a method to the Association class
that checks to see if an association's job would put them over their
max_nodes limit under a particular queue.