database: add duration columns to the "jobs" table
Problem: The requested duration and actual duration for jobs submitted
by associations are stored in and calculated from extracting values from
the jobspec, but they are not conveniently stored anywhere for use
throughout flux-accounting. It would be nice if these types of values
for jobs were easily avaialable for tracking on a per-association or
other basis.

Add two new columns to the jobs table: requested_duration and
actual_duration, floating-point values that store the job duration
requested by the user and the job's actual duration, respectively.
Update the schema version of the flux-accounting DB to account for the
addition of two new columns to the "jobs" table.