accounting: add functions to unpack JSON DB info
Problem: In order to initialize the data structures in the priority
plugin with the JSON object provided at load time, there need to be
functions to unpack the key-value pairs within that JSON object and
place them in the appropriate data structures.

Add functions to accounting.hpp and accounting.cpp to unpack the
key-value pairs provided in the JSON object passed in at plugin load
time and place the values in their respective data structures in the
priority plugin.