plugin: update job, resource counts on reload
Problem: The priority plugin relies on keeping an association's active
job and resource counts by associating an aux item with each job. This
aux item gets removed when the plugin is unloaded and has to be re-added
when the plugin is loaded back in, and the flux-accounting database
information isn't always immediately loaded when the plugin itself is
loaded; if this is the case, the plugin sets a temporary aux item on the
job while it waits for information from the database. Now, the plugin
accepts an optional key-value argument during load time that can
initialize the plugin with database information, which no longer
requires the use of a temporary aux item to be associated with the job,
and can instead directly track an association's job and resource counts
when it rediscovers active jobs in job.new.

Add a check for the RUN job state of a job in job.new. If this is
discovered, that means the plugin was reloaded while this job was
running, so increment the active, running, and resource counts of the
association that submitted this job.