plugin: set load order of active jobs during init
Problem: The order in which jobs are seen when the plugin is loaded is
not guaranteed *unless* a sort mode is passed when initializing the
plugin (see flux-jobtap-plugins(7)), which is required in order to
accurately track active jobs and their flux-accounting dependencies
after the priority plugin is reloaded.

Unpack an optional "sort" key when loading the plugin, and if unpacked,
call flux_jobtap_set_load_sort_order () to see jobs by reverse state in
order to accurately track and enforce limits and job/resource counts
per-association and per-queue.